[refactor] #29 - admincontroller에서 참조하던 ssecontroller제거 후 service에서 참조
